Save two bytes of memory, let's not branch on undefined, fix warning.
git-svn-id: https://svn.zsnes.com/rep/zsnes/trunk@3809 8cdf4848-9e13-0410-98e6-c9478ff1fb58

+++ src/gui/guifuncs.c
@@ -1685,7 +1685,7 @@ void GUILoadKeysJumpTo()
     }
   }
 
-  if (found < (unsigned int)*entries)
+  if (found < *entries)
   {
     *currentviewloc = found;
     *currentcursloc = found;
